Correctly analyse accessed rule variables in the presence of grammar cycles
If there are grammar cycles, the computed `accessedParams` and `refScope`
could be incorrect.  For `accessedParams`, when a rule reference is
encountered for the second time as part of a cycle, its `accessedParams`
will be empty.  That will get incorrectly propagated into rules
participating in the cycle.  As a result, the rules involved will not
pass down the parameter value and will instead reinitialize it when
the cyclic rule is encountered.  To fix this issue, we iterate the
analysis until we arrive at a fixed point.

Simiarly, if the first reference to a rule happens to have an assignment,
the refScope will be nulled out when the referenced rule is traversed.
Since traversal never visits the same node twice, a later reference
which does not contain an assignment will not revisit the rule and
the refScope won't be correctly recorded.  To fix this issue, we
separate out the `isAssigned` analysis from the `refScope` analysis,
and ensure that we don't traverse into a rule if `refScope` would be
null.  This ensures that a later reference with a non-null refScope
won't find the rule "already visited".

Change-Id: I969106c536b0ebcf56b0f4b464030dfd179a998f